backend/requirements.txt
django
djangorestframework
django-cors-headers
celery
redis
openai
langchain
langchain-openai
langgraph
requests
Pillow
python-dotenv
gunicorn
